+++ src/components/layout/list/Item.stories.tsx
@@ -1,11 +1,21 @@
 import React from "react";
 
+import { omit } from "@ezez/utils";
+
 import type { StoryObj, Meta } from "@storybook/react";
 
 import { NextLink } from "../../../demo/NextLink";
 
 import { Item } from "./Item";
 import { List } from "./List";
+import { Label } from "./Label";
+import { Value } from "./Value";
+
+const demoControl = {
+    table: {
+        category: "Demo controls",
+    },
+};
 
 const meta: Meta = {
     title: "Components/Layout/List/Item",
@@ -15,8 +25,16 @@ const meta: Meta = {
         inset: {
             type: "boolean",
         },
+        selected: {
+            type: "boolean",
+        },
         Link: {
             controls: "none",
+        },
+        renderList: {
+            ...demoControl,
+            type: "boolean",
+            name: "Render <List> around",
         },
     },
 };
@@ -26,11 +44,28 @@ type Story = StoryObj<typeof Item>;
 const Primary: Story = {
     args: {
         children: "I am a list item",
-        inset: false,
         Link: NextLink,
     },
-    render: (args) => {
-        return <List><Item {...args} /></List>;
+    render: (rawProps) => {
+        const renderList = (rawProps as { renderList: boolean | undefined }).renderList;
+        // @ts-expect-error Storybook is crazy
+        const args = omit(rawProps, ["renderList"]);
+
+        const items = [
+            <Item key={1} {...args} />,
+            <Item key={2}>Other item</Item>,
+            <Item key={3}>Other item</Item>,
+        ];
+
+        if (renderList) {
+            return (
+                <List>{items}</List>
+            );
+        }
+
+        return (
+            <div>{items}</div>
+        );
     },
 };
 
@@ -44,7 +79,13 @@ const Alignment: Story = {
         ratio: "1/2/1",
     },
     render: (args) => {
-        return <List><Item {...args} /></List>;
+        return (
+            <List>
+                <Item {...args} />
+                <Item>Other item</Item>
+                <Item>Other item</Item>
+            </List>
+        );
     },
 };
 
@@ -56,7 +97,32 @@ const OnClick: Story = {
         },
     },
     render: (args) => {
-        return <List><Item {...args} /></List>;
+        return (
+            <List>
+                <Item {...args} />
+                <Item>Other item</Item>
+                <Item>Other item</Item>
+            </List>
+        );
+    },
+};
+
+const Complex: Story = {
+    args: {
+        children: [
+            <Label key={"l"}>Hi</Label>,
+            <Value key={"v"}>Value</Value>,
+        ],
+        selected: true,
+    },
+    render: (args) => {
+        return (
+            <List>
+                <Item {...args} />
+                <Item>Other item</Item>
+                <Item>Other item</Item>
+            </List>
+        );
     },
 };
 
@@ -64,6 +130,7 @@ export {
     Primary,
     Alignment,
     OnClick,
+    Complex,
 };
 
 export default meta;